backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
langchain==0.1.1
langchain-openai==0.0.5
langchain-huggingface==0.0.1
langchain-community==0.0.10
faiss-cpu==1.7.4
huggingface-hub==0.19.4
sentence-transformers==2.2.2
transformers==4.35.2
APScheduler==3.10.4
requests==2.31.0
Werkzeug==3.0.1
bcrypt==4.1.1
